breakout mod Dither adaptor iterator
